requirements.txt
aiogram==2.25.1
librosa==0.10.0
soundfile==0.12.1
SpeechRecognition==3.10.0
python-dotenv